util/pixel.c: Use 'intptr_t' instead of 'long' for pointer casing (harmless).
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@62809 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/GuestHost/OpenGL/util/pixel.c
@@ -1554,7 +1554,7 @@ void crPixelCopy2D( GLsizei width, GLsizei height,
 
         /* handle the alignment */
         if (srcPacking->alignment != 1) {
-            i = ((long) src) % srcPacking->alignment;
+            i = ((intptr_t) src) % srcPacking->alignment;
             if (i)
                 src += srcPacking->alignment - i;
             i = (long) srcRowStrideBytes % srcPacking->alignment;
@@ -1563,7 +1563,7 @@ void crPixelCopy2D( GLsizei width, GLsizei height,
         }
 
         if (dstPacking->alignment != 1) {
-            i = ((long) dst) % dstPacking->alignment;
+            i = ((intptr_t) dst) % dstPacking->alignment;
             if (i)
                 dst += dstPacking->alignment - i;
             i = (long) dstRowStrideBytes % dstPacking->alignment;
